Refactor client impls
